- Improved SET_motor.ino - it appears that motor_ctrl:ROM is increasing arbitrarily to 144. Trying to debug.

+++ SET_motor/interface.h
@@ -7,17 +7,17 @@ int getROM();
 float getpressureThreshold();
 float getStepSize();
 
-int getBPM();
-int getVOL();
-int getIE();
+float getBPM();
+float getVOL();
+float getIE();
 int getASIST();
 int getALARM();
 float getSwitchToBag();
 float getBagToCentre();
 
-void setBPM(int bpm);
-void setVOL(int vol);
-void setIE(int ie);
+void setBPM(float bpm);
+void setVOL(float vol);
+void setIE(float ie);
 void setASIST(int assist);
 void setALARM(int alarm);
 void setSwitchToBag(float switch_to_bag);
